✏️ Implements 'read input' command `,`
Finishes the implementation of `Evaluator.handleInputByte` reading the head of
`vm.input` into the cells of the `VirtualMachine`. Adds a `DATA` section to the
UI in which the user can input the data to be given to the program entered in
the `SOURCE` section. Need to add some documentation to the UI on how brainfuck
and how to properly use the interface.

Also fixes an issue with the tape size being too short: it should be at least
30_000 cells. However, due to technical constraints having an array with a
length of 1000+ causes runtime issues, so instead a `Cells` type is created in
the `VirtualMachine` module that implements the 30_000 across a 2D array of 1000
rows each having 30 columns. The `Evaluator` then uses these functions to
manipulate the cells of the `VirtualMachine`.

Adds the `Brainfuck.Program` module which contains a range of reference
BF programs from brainfuck.org used for testing the implementation. Need to make
authorship clear of these programs and also incorporate them into the testing
module.