✏️ Adds navigation, reference manual page and source archives page
Adds the `Page` type to toggle between the reference manual, source archives and
interpreter pages of the web app.

Adds the reference manual page containing an old school description of the
BF-4000 (Brainfuck) machine language, its features, syntax, and semantics. This
page is purely informational with no actions to be performed by the user.

Adds the source archives page containing a list of (mocked) brainfuck files
for the user to select between.