✏️ Swaps out the quick-n-dirty lexer and parser with a parser combinator
Lexer:
Deletes the `Lexer` as it is not really needed for such a simply grammar.

Parser:
Changes the existing `Parser` logic to instead use the Elm parser combinator
library allowing for a more declarative parsing logic.

Filters out all symbols not in the brainfuck language rather than turn them into
comments, which could be improved.

Also worth considering if the AST should have an explicit `Block` vs. `Command`
in its type constructor to make the `case` slightly cleaner in the evaluate
functions.

Evaluator:
Turns the evaluation of a program into a `List.stopableFoldl` allowing the
program to exit early if an error occurs.

Main:
Adds a `ClearOutput` message to trigger from the UI.

Needs to turn a failed parsing into a proper error to show to the user.

ParserTests:
Adds initial tests for the `Parser` module.