✏️ Rewrites evaluator to do stepwise evaluation
Rewrites the Evaluator to carry an `EvaluationContext` around instead of just
the `VirtualMachine`. This context includes the state of the evaluation and
breadcrumbs (pointer) to the current evaluation step. With this, we are then
able to turn the evaluation into having in discrete steps, which in turn allows
us to sync the steps with Elm's runtime giving us a time-traveling debugger for
free thanks to Elm's built-in tooling.